Persistence has been added to the app through localStorage.
All the projects info will be saved to the project
array to be used when revisiting the app.
All the tasks under each project has been saved to
the tasks array to be used when revisiting the app
And finally the info in the notes has also been
saved to localStorage for use when revisiting the
app.